% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_write_save.R
\name{fcs_write_save}
\alias{fcs_write_save}
\title{Write out the unmixed data and save as RDS}
\usage{
fcs_write_save(unmixed_data, sample_name)
}
\arguments{
\item{unmixed_data}{Matrix, the unmixed full-stained data}

\item{sample_name}{string, name of your sample}
}
\value{
return the fcs files and saved RDS
}
\description{
Take the matrix of the unmixed data, write to FCS files
}
