% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_write_save_list.R
\name{fcs_write_save_list}
\alias{fcs_write_save_list}
\title{Write out to fcs for a list of unmixed data and save as RDS}
\usage{
fcs_write_save_list(unmixed_data_list, rds_name = NULL)
}
\arguments{
\item{unmixed_data_list}{List of matrices, the unmixed data}

\item{rds_name}{String, a name to save the list of unmixed data matrices as an .rds}
}
\value{
return the fcs files and saved RDS
}
\description{
Take the list of matrices of the unmixed data, write to several FCS files, and save RDS.
}
