% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference.R
\name{load_reference}
\alias{load_reference}
\title{Generate spectral signatures}
\usage{
load_reference(
  ref_dir,
  af = TRUE,
  af_dir,
  change_order = FALSE,
  custom_order = FALSE,
  channel_order = NULL
)
}
\arguments{
\item{ref_dir}{Directory containing reference control .fcs, string}

\item{af}{Whether to do autofluorescence extraction. Default = TRUE, recommended.}

\item{af_dir}{Directory containing AF .fcs, string}

\item{change_order}{Logical, only use when exporting gated .fcs files from flowjo. Default order is as in Aurora in WIMR, Australia.}

\item{custom_order}{Change order to another Aurora's channels.}

\item{channel_order}{When custom_order = TRUE, enter a string vector of the Aurora channels.}
}
\value{
return a data.frame of channels, normalised signatures, marker names
}
\description{
Load reference controls and unstained controls and generate signatures
}
