% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_generate.R
\name{M_generate}
\alias{M_generate}
\title{A function to generate spectral signature matrix}
\usage{
M_generate(df, refs, plot = TRUE)
}
\arguments{
\item{df}{data.frame, the signature data frame, for plotting}

\item{refs}{A list of reference controls & unstained control}

\item{plot}{Logical, whether a plot of spectral signatures should be generated}
}
\value{
return a spectral signature matrix M
}
\description{
A function to generate spectral signature matrix
}
