% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_diagnose.R
\name{panel_diagnose}
\alias{panel_diagnose}
\title{Build and visualise a few pre-unmixing diagnostics}
\usage{
panel_diagnose(M, plot = TRUE)
}
\arguments{
\item{M}{Matrix, the spectral signature matrix}

\item{plot}{Logical. Default TRUE. Whether plots should be generated.}
}
\value{
return the covariance matrix, cosine similarity matrix, and condition number
}
\description{
Take the spectral signature matrix, preview covariance matrix, cosine similarity matrix, and the condition number.
}
