% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_OLS.R
\name{unmix_OLS}
\alias{unmix_OLS}
\title{Unmixing OLS, standard solution}
\usage{
unmix_OLS(
  M,
  fs_raw,
  aurora_channel_change = FALSE,
  channels = NULL,
  plot = TRUE
)
}
\arguments{
\item{M}{Matrix, the signature matrix}

\item{fs_raw}{string, the name of the raw full-stained .fcs file}

\item{aurora_channel_change}{Logical, Whether you would like to change Aurora configuration. Default to WIMR, Australia.}

\item{channels}{If aurora_channel_change is TRUE, enter a combined string of names of channels}

\item{plot}{Plotting the unmixing matrix. Default TRUE.}
}
\value{
return a matrix of unmixed data
}
\description{
Take the spectral signature M and unmix the full-stained sample via closed-form solution of OLS.
}
