% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_OLS_SVD.R
\name{unmix_OLS_SVD}
\alias{unmix_OLS_SVD}
\title{Unmixing OLS, M+ via SVD method}
\usage{
unmix_OLS_SVD(
  M,
  fs_raw,
  aurora_channel_change = FALSE,
  channels = NULL,
  plot = TRUE
)
}
\arguments{
\item{M}{Matrix, the signature matrix}

\item{fs_raw}{string, the name of the raw full-stained .fcs file}

\item{aurora_channel_change}{Logical, Whether you would like to change Aurora configuration. Default to WIMR, Australia.}

\item{channels}{If aurora_channel_change is TRUE, enter a combined string of names of channels}

\item{plot}{Plotting the unmixing matrix via SVD. Default TRUE.}
}
\value{
return a matrix of unmixed data
}
\description{
Take the spectral signature M and solve for an unmixing matrix via SVD, then unmix the full-stained sample.
}
