% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_WLS.R
\name{unmix_WLS}
\alias{unmix_WLS}
\title{Unmixing Weighted Least-Squares, standard solution}
\usage{
unmix_WLS(M, fs_raw, aurora_channel_change = FALSE, channels = NULL)
}
\arguments{
\item{M}{Matrix, the signature matrix}

\item{fs_raw}{string, the name of the raw full-stained .fcs file}

\item{aurora_channel_change}{Logical, Whether you would like to change Aurora configuration. Default to WIMR, Australia.}

\item{channels}{If aurora_channel_change is TRUE, enter a combined string of names of channels}
}
\value{
return a matrix of unmixed data
}
\description{
Take the spectral signature M, calculate the weights by 1/detector MFI, and unmix the full-stained sample via closed-form solution.
}
