% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_WLS_list.R
\name{unmix_WLS_list}
\alias{unmix_WLS_list}
\title{Unmixing WLS, standard solution for a list of samples}
\usage{
unmix_WLS_list(M, fs_raw_dir, aurora_channel_change = FALSE, channels = NULL)
}
\arguments{
\item{M}{Matrix, the signature matrix}

\item{fs_raw_dir}{a string of the path that stores a list of fcs files to unmix.}

\item{aurora_channel_change}{Logical, Whether you would like to change Aurora configuration. Default to WIMR, Australia.}

\item{channels}{If aurora_channel_change is TRUE, enter a combined string of names of channels}
}
\value{
return a list of matrices of unmixed data
}
\description{
Take the spectral signature M and unmix a list of full-stained samples via closed-form solution of OLS.
}
